feat: move walletConnect store clearing logic to disconnect function
