+++ account-kit/core/src/actions/disconnect.ts
@@ -29,6 +29,9 @@ export async function disconnect(config: AlchemyAccountsConfig): Promise<void> {
 
   config.store.persist.clearStorage();
 
+  // Clear Wallet Connect store
+  clearWalletConnectStore();
+
   config.store.setState(() => config.store.getInitialState());
 
   config.store.setState((state) => ({
@@ -38,3 +41,25 @@ export async function disconnect(config: AlchemyAccountsConfig): Promise<void> {
     ),
   }));
 }
+
+// Function to clear the Wallet Connect store to prevent
+// Persistence of Wallet Connect connection state on error.
+const clearWalletConnectStore = () => {
+  // Open Wallet Connect Indexed DB
+  const dbOpenRequest = indexedDB.open("WALLET_CONNECT_V2_INDEXED_DB");
+
+  dbOpenRequest.onsuccess = () => {
+    const db = dbOpenRequest.result;
+
+    const txn = db.transaction(["keyvaluestorage"], "readwrite");
+
+    const store = txn.objectStore("keyvaluestorage");
+
+    // Clear Store
+    store.clear();
+  };
+
+  dbOpenRequest.onerror = () => {
+    console.error("Error opening Wallet Connect DB. Cannot clear store.");
+  };
+};
